Renamed aggregator files and base class
